server/pyproject.toml
[tool.poetry]
name = "fennec-api"
version = "0.1.0"
description = ""
authors = ["saiqi <julien.bernard.iphone@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fennec_api"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.6"
aiofiles = "^23.1.0"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic-settings = "^2.2.1"
uvicorn = "^0.27.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.6.3"}
python-multipart = "^0.0.9"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
mypy = "^1.8"
pytest-asyncio = "^0.21.0"
ruff = "^0.3.0"
types-passlib = "^1.7.7.20240106"
aiosqlite = "^0.20.0"
types-python-jose = "^3.3.4.20240106"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["pydantic.mypy"]

follow_imports = "silent"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[[tool.mypy.overrides]]
module = ["aiofiles"]
ignore_missing_imports = true

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402"]
"**/{tests,docs,tools}/*" = ["E402"]
"**/migrations/env.py" = ["E402","F401"]
